+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/gtceu/ControllablePeripheralPlugin.kt
@@ -1,6 +1,7 @@
 package site.siredvin.peripheralworks.integrations.gtceu
 
 import com.gregtechceu.gtceu.api.capability.IControllable
+import com.gregtechceu.gtceu.integration.cctweaked.peripherals.ControllablePeripheral
 import dan200.computercraft.api.lua.LuaFunction
 import dan200.computercraft.api.lua.MethodResult
 import site.siredvin.tweakium.modules.peripheral.api.IPeripheralPlugin
@@ -10,16 +11,19 @@ class ControllablePeripheralPlugin(private val controllable: IControllable) : IP
         const val TYPE = "gtceu:controllable"
     }
 
+    override val additionalType: String
+        get() = TYPE
+
     @LuaFunction(mainThread = true)
-    fun isWorkingEnabled(): MethodResult = MethodResult.of(controllable.isWorkingEnabled)
+    fun isWorkingEnabled(): MethodResult = ControllablePeripheral.isWorkingEnabled(controllable)
 
     @LuaFunction(mainThread = true)
     fun setWorkingEnabled(enabled: Boolean) {
-        controllable.isWorkingEnabled = enabled
+        ControllablePeripheral.setWorkingEnabled(controllable, enabled)
     }
 
     @LuaFunction(mainThread = true)
     fun setSuspendAfterFinish(suspendAfterFinish: Boolean) {
-        controllable.setSuspendAfterFinish(suspendAfterFinish)
+        ControllablePeripheral.setSuspendAfterFinish(controllable, suspendAfterFinish)
     }
 }